// src/app/api/model/roleW.ts
/**
 * GestionRFQ API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */


export type RoleW = 0 | 1;

export const RoleW = {
    NUMBER_0: 0 as RoleW,
    NUMBER_1: 1 as RoleW
};