// src/app/api/model/workerDto.ts
/**
 * GestionRFQ API
 *
 * 
 *
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
import { RoleW } from './roleW';


export interface WorkerDto { 
    id?: number;
    nom?: string | null;
    role?: RoleW;
}
export namespace WorkerDto {
}